Refactored folder structure, switch to json results and data converting
