+++ src/Equinor.ProCoSys.Config/ProcosysJsFrontend/ConfigurationSet.cs
@@ -1,10 +1,10 @@
 ﻿using System.Collections.Generic;
 
-namespace Equinor.ProCoSys.Config
+namespace Equinor.ProCoSys.Config.ProcosysJsFrontend
 {
     public class ConfigurationSet
     {
-        public Dictionary<string, string> Configuration { get; } = new Dictionary<string, string>();
+        public Dictionary<string, object> Configuration { get; } = new Dictionary<string, object>();
         public Dictionary<string, bool> FeatureFlags { get; } = new Dictionary<string, bool>();
     }
 }
